refactor: move arch virtmem init routine into arch module
